+++ crowdfund/settings.py
@@ -31,9 +31,9 @@ ALLOWED_HOSTS = []
 # Application definition
 
 AUTH_USER_MODEL = 'accounts.CustomUser'
-LOGIN_URL = '/' # All login redirects point to the landing page
+LOGIN_URL = 'accounts:login' # All login redirects point to the login page
 LOGIN_REDIRECT_URL = '/app/' # Fallback redirect for authenticated users
-LOGOUT_REDIRECT_URL = '/' # Redirect to landing page after logout
+LOGOUT_REDIRECT_URL = 'home' # Redirect to landing page after logout
 
 INSTALLED_APPS = [
     'django.contrib.admin',